refactor: fold soft rw cancel tests, align wrappers
Address review on the AsyncSoftReadWriteLock cancellation fix.

Fold the cancellation tests into tests/soft_rw/test_soft_rw_async.py and
drop the separate file, reusing the module's _make helper.

Collapse the duplicate except branch in AsyncReadWriteLock._run_acquire now
that _raise_cancelled_error is shared, so both async wrappers read the same.

Tighten the changelog and the reconciliation comments.
